feat: Add `authMiddleware` Nextjs Middleware util to protect directories
This adds a new utility function `authMiddleware` which leverages Next v12's [Middleware](https://nextjs.org/docs/middleware) feature to add blanket protection for all routes in a given directory via a `_middleware` file.

It is based on `getUser`, which will attempt to recover a user's session via their `access_token` or `refresh_token`. If a user's session is verified to exist, the incoming request is forwarded. Otherwise, they will be redirected to the given route, defaulting the the site root.

A usage example was also added to the README